Remove duplications to pass quality gate: Refactor image picker logic by centralizing functionality into `useImagePicker` hook and updating forms accordingly
